Move heroes under games module
